docs: Add biological foundations section referencing Gordon (1996)
- Add "Biological Foundations" section to theory.md explaining how
  SWARM's distributed governance draws on behavioral ecology research
- Map ant colony task allocation mechanisms to SWARM concepts:
  - Threshold model → Circuit breakers
  - Interaction-based switching → Quality signal responses
  - Stigmergy → Event logs and shared state
- Document adversarial parallels (death spirals, parasitic exploitation)
- Add Gordon (1996) and related swarm intelligence references
- Add cross-reference from governance.md to biological foundations

+++ swarm/metrics/time_horizons.py
@@ -14,6 +14,7 @@ across different task durations in multi-agent simulations.
 
 from dataclasses import dataclass, field
 from typing import Any
+
 import numpy as np
 
 
@@ -188,18 +189,18 @@ class AgentCapabilityProfile:
             return self.base_reliability_10min
 
         # Log-scale decay from 10-minute baseline
-        log_ratio = np.log10(max(minutes, 1) / 10)
-        decay = self.horizon_decay_rate * max(0, log_ratio)
+        log_ratio = float(np.log10(max(minutes, 1) / 10))
+        decay = self.horizon_decay_rate * max(0.0, log_ratio)
 
         # Apply capability scaling
         base = self.base_reliability_10min * self.capability_level
-        return max(0.0, min(1.0, base - decay))
+        return float(max(0.0, min(1.0, base - decay)))
 
     def expected_quality(self, minutes: int) -> float:
         """Expected output quality given task duration."""
         reliability = self.reliability_at_horizon(minutes)
         # Quality degrades faster than reliability
-        return reliability ** 1.2
+        return float(reliability ** 1.2)
 
     def compute_cost(self, minutes: int) -> float:
         """Total compute cost for a task of given duration."""
